+++ catalog/app/containers/Bucket/PackageTree/RevisionInfo.tsx
@@ -3,22 +3,16 @@ import * as dateFns from 'date-fns'
 import * as R from 'ramda'
 import * as React from 'react'
 import { Link as RRLink } from 'react-router-dom'
+import type { ResultOf } from '@graphql-typed-document-node/core'
 import * as M from '@material-ui/core'
 
 import * as Notifications from 'containers/Notifications'
-import * as APIConnector from 'utils/APIConnector'
-import AsyncResult from 'utils/AsyncResult'
-import { useData } from 'utils/Data'
 import * as NamedRoutes from 'utils/NamedRoutes'
 import { linkStyle } from 'utils/StyledLink'
 import copyToClipboard from 'utils/clipboard'
+import { UseQueryResult } from 'utils/useQuery'
 
-import * as requests from './requests'
-
-function useRevisionsData({ bucket, name }) {
-  const req = APIConnector.use()
-  return useData(requests.getPackageRevisions, { req, bucket, name, perPage: 5 })
-}
+import REVISION_LIST_QUERY from './gql/RevisionList.generated'
 
 const useRevisionInfoStyles = M.makeStyles((t) => ({
   revision: {
@@ -27,6 +21,7 @@ const useRevisionInfoStyles = M.makeStyles((t) => ({
     display: 'inline-flex',
   },
   mono: {
+    // @ts-expect-error
     fontFamily: t.typography.monospace.fontFamily,
   },
   line: {
@@ -43,37 +38,43 @@ const useRevisionInfoStyles = M.makeStyles((t) => ({
   },
 }))
 
-export default function RevisionInfo({ revisionData, revision, bucket, name, path }) {
+interface RevisionInfoProps {
+  bucket: string
+  name: string
+  path: string
+  hashOrTag: string
+  hash?: string
+  revisionListQuery: UseQueryResult<ResultOf<typeof REVISION_LIST_QUERY>>
+}
+
+export default function RevisionInfo({
+  bucket,
+  name,
+  hash,
+  hashOrTag,
+  path,
+  revisionListQuery,
+}: RevisionInfoProps) {
   const { urls } = NamedRoutes.use()
   const { push } = Notifications.use()
   const classes = useRevisionInfoStyles()
 
-  const listRef = React.useRef()
-  const [anchor, setAnchor] = React.useState()
+  const listRef = React.useRef<HTMLUListElement>(null)
+  const [anchor, setAnchor] = React.useState<HTMLSpanElement | null>(null)
   const [opened, setOpened] = React.useState(false)
   const open = React.useCallback(() => setOpened(true), [])
   const close = React.useCallback(() => setOpened(false), [])
 
-  const revisionsData = useRevisionsData({ bucket, name })
-  const data = revisionsData.case({
-    Ok: (revisions) =>
-      revisionData.case({
-        Ok: ({ hash }) =>
-          AsyncResult.Ok(revisions.map((r) => ({ ...r, selected: r.hash === hash }))),
-        Err: () => AsyncResult.Ok(revisions),
-        _: R.identity,
-      }),
-    _: R.identity,
-  })
+  const getHttpsUri = (h: string) =>
+    `${window.origin}${urls.bucketPackageTree(bucket, name, h, path)}`
 
-  const getHttpsUri = (r) =>
-    `${window.origin}${urls.bucketPackageTree(bucket, name, r.hash, path)}`
-
-  const copyHttpsUri = (r, containerRef) => (e) => {
-    e.preventDefault()
-    copyToClipboard(getHttpsUri(r), { container: containerRef && containerRef.current })
-    push('Canonical URI copied to clipboard')
-  }
+  const copyHttpsUri =
+    (h: string, containerRef?: React.RefObject<HTMLUListElement>) =>
+    (e: React.MouseEvent) => {
+      e.preventDefault()
+      copyToClipboard(getHttpsUri(h), { container: containerRef?.current || undefined })
+      push('Canonical URI copied to clipboard')
+    }
 
   return (
     <>
@@ -82,25 +83,22 @@ export default function RevisionInfo({ revisionData, revision, bucket, name, pat
         className={classes.revision}
         onClick={open}
         ref={setAnchor}
-        title={revision.length > 10 ? revision : undefined}
+        title={hashOrTag.length > 10 ? hashOrTag : undefined}
       >
-        {R.take(10, revision)} <M.Icon>expand_more</M.Icon>
+        {R.take(10, hashOrTag)} <M.Icon>expand_more</M.Icon>
       </span>
 
-      {revisionData.case({
-        Ok: (r) => (
-          <M.IconButton
-            size="small"
-            title="Copy package revision's canonical catalog URI to the clipboard"
-            href={getHttpsUri(r)}
-            onClick={copyHttpsUri(r)}
-            style={{ marginTop: -4, marginBottom: -4 }}
-          >
-            <M.Icon>link</M.Icon>
-          </M.IconButton>
-        ),
-        _: () => null,
-      })}
+      {!!hash && (
+        <M.IconButton
+          size="small"
+          title="Copy package revision's canonical catalog URI to the clipboard"
+          href={getHttpsUri(hash)}
+          onClick={copyHttpsUri(hash)}
+          style={{ marginTop: -4, marginBottom: -4 }}
+        >
+          <M.Icon>link</M.Icon>
+        </M.IconButton>
+      )}
 
       <M.Popover
         open={opened && !!anchor}
@@ -110,24 +108,15 @@ export default function RevisionInfo({ revisionData, revision, bucket, name, pat
         transformOrigin={{ vertical: 'top', horizontal: 'center' }}
       >
         <M.List className={classes.list} ref={listRef}>
-          {AsyncResult.case(
-            {
-              Ok: R.ifElse(
-                R.isEmpty,
-                () => (
-                  <M.ListItem>
-                    <M.ListItemText
-                      primary="No revisions found"
-                      secondary="Looks like this package has been deleted"
-                    />
-                  </M.ListItem>
-                ),
-                R.map((r) => (
+          {revisionListQuery.case({
+            data: (d) =>
+              d.package ? (
+                d.package.revisions.page.map((r) => (
                   <M.ListItem
-                    key={r.hash}
+                    key={`${r.hash}:${r.modified.valueOf()}`}
                     button
                     onClick={close}
-                    selected={r.selected}
+                    selected={r.hash === hash}
                     component={RRLink}
                     to={urls.bucketPackageTree(bucket, name, r.hash, path)}
                   >
@@ -146,34 +135,39 @@ export default function RevisionInfo({ revisionData, revision, bucket, name, pat
                     <M.ListItemSecondaryAction>
                       <M.IconButton
                         title="Copy package revision's canonical catalog URI to the clipboard"
-                        href={getHttpsUri(r)}
-                        onClick={copyHttpsUri(r, listRef)}
+                        href={getHttpsUri(r.hash)}
+                        onClick={copyHttpsUri(r.hash, listRef)}
                       >
                         <M.Icon>link</M.Icon>
                       </M.IconButton>
                     </M.ListItemSecondaryAction>
                   </M.ListItem>
-                )),
-              ),
-              Err: () => (
+                ))
+              ) : (
                 <M.ListItem>
-                  <M.ListItemIcon>
-                    <M.Icon>error</M.Icon>
-                  </M.ListItemIcon>
-                  <M.Typography variant="body1">Error fetching revisions</M.Typography>
+                  <M.ListItemText
+                    primary="No revisions found"
+                    secondary="Looks like this package has been deleted"
+                  />
                 </M.ListItem>
               ),
-              _: () => (
-                <M.ListItem>
-                  <M.ListItemIcon>
-                    <M.CircularProgress size={24} />
-                  </M.ListItemIcon>
-                  <M.Typography variant="body1">Fetching revisions</M.Typography>
-                </M.ListItem>
-              ),
-            },
-            data,
-          )}
+            error: () => (
+              <M.ListItem>
+                <M.ListItemIcon>
+                  <M.Icon>error</M.Icon>
+                </M.ListItemIcon>
+                <M.Typography variant="body1">Error fetching revisions</M.Typography>
+              </M.ListItem>
+            ),
+            fetching: () => (
+              <M.ListItem>
+                <M.ListItemIcon>
+                  <M.CircularProgress size={24} />
+                </M.ListItemIcon>
+                <M.Typography variant="body1">Fetching revisions</M.Typography>
+              </M.ListItem>
+            ),
+          })}
           <M.Divider />
           <M.ListItem
             button